requirements.txt
blend_modes==2.1.0
Flask==1.1.2
numpy==1.19.4
Pillow==9.0.0
requests==2.25.0
tweepy==4.4.0
